+++ src/AttendanceChecker/GUI/Controllers/StudentViewController.java
@@ -55,7 +55,7 @@ public class StudentViewController implements Initializable {
             absencePieChart.setTitle("Absence chart:");
             selectedStudentMostAbsentDay.setText(studentModel.getValueFromDay(selectedStudent));
             selectedStudentTotalAbsenceDays.setText(selectedStudent.getAbsentDays() + "");
-            selectedStudentPercentageAbsence.setText((selectedStudent.getAbsentDays()/studentModel.getTotalDays()) * 100 + "%");
+            selectedStudentPercentageAbsence.setText(((double)selectedStudent.getAbsentDays()/(double)studentModel.getTotalDays()) * 100 + "%");
         } catch (IOException | SQLException e) {
             e.printStackTrace();
         }
